Final app

+++ app/api/posts/[id]/route.tsx
@@ -1,32 +1,52 @@
-import { NextApiRequest } from 'next';
-
 import { PrismaClient } from '@prisma/client'
-import { authenticate } from '@/lib/stytch-client';
+import { authorize } from '@/lib/stytch-client';
 import { NextRequest } from 'next/server';
 
 const prisma = new PrismaClient()
 
-export async function GET(request: NextApiRequest, { params }: { params: { id: string }}) {
-    const { organization } = await authenticate();
+export async function GET(request: NextRequest, { params }: { params: { id: string }}) {
+    const organization_id = request.nextUrl.searchParams.get("organization_id");
+
+    if(!organization_id) return Response.json({ success: false, message: 'Organization ID is required' }, { status: 400 });
+
+    const { authorized } = await authorize({
+        organization_id,
+        resource_id: 'post',
+        action: 'read',
+    });
+
+    if(!authorized) return Response.json({ success: false, message: 'Unauthorized', post: null }, { status: 403 });
+
     const post = await prisma.post.findUnique({
         where: {
             id: Number(params.id),
-            organization: organization.organization_id
+            organization: organization_id
         }
     })
-    if(!post) return Response.json({ success: false, message: 'Post not found' }, { status: 404 });
-    console.log(post);
+    if(!post) return Response.json({ success: false, message: 'Post not found', post: null }, { status: 404 });
+
     return Response.json({ success: true, post });
 }
 
 export async function PUT(request: NextRequest, { params }: { params: { id: string }}) {
-    const { organization } = await authenticate();
+    const organization_id = request.nextUrl.searchParams.get("organization_id");
+
+    if(!organization_id) return Response.json({ success: false, message: 'Organization ID is required' }, { status: 400 });
+
+    const { authorized } = await authorize({
+        organization_id,
+        resource_id: 'post',
+        action: 'update',
+    });
+
+    if(!authorized) return Response.json({ success: false, message: 'Unauthorized', post: null }, { status: 403 });
+
     const data: { title: string, content: string } = await request.json();
     
         const post = await prisma.post.update({
             where: {
                 id: Number(params.id),
-                organization: organization.organization_id
+                organization: organization_id
             },
             data: {
                 title: data.title,
@@ -37,12 +57,23 @@ export async function PUT(request: NextRequest, { params }: { params: { id: stri
         return Response.json({ success: true, post });
 }
 
-export async function DELETE(request: NextApiRequest, { params }: { params: { id: string }}) {
-    const { organization } = await authenticate();
+export async function DELETE(request: NextRequest, { params }: { params: { id: string }}) {
+    const organization_id = request.nextUrl.searchParams.get("organization_id");
+
+    if(!organization_id) return Response.json({ success: false, message: 'Organization ID is required' }, { status: 400 });
+
+    const { authorized } = await authorize({
+        organization_id,
+        resource_id: 'post',
+        action: 'delete',
+    });
+
+    if(!authorized) return Response.json({ success: false, message: 'Unauthorized', post: null }, { status: 403 });
+
     const post = await prisma.post.delete({
         where: {
             id: Number(params.id),
-            organization: organization.organization_id
+            organization: organization_id
         }
     });
     return Response.json({ success: true, post });
